Server/Client refactoring part 1 [see details] - BUILD FAILS
This is a developing commit, the build fails and it isn't the full implementation.
The server has createUser and userLogin methods which will communicate with both rmi and socket.

A ServerSocketConnectionHandler object is responsible to handle a single socket connection, same for rmi (future impl.). A ServerSocketHandler (rmi too) will represent a socket (rmi) server listening for incoming connections. Both connectionHandlers will be subclassed by a ServerConnectionHandler which will expose methods (called by the server) and hide the socket/rmi implementation.

That's all folks! (I think)
for other info: @rofonafets on telegram